e2e_tests/test_visual_scheduler_performance.py
"""Performance tests for Visual Scheduler with large datasets."""
import pytest
from playwright.sync_api import Page, expect
from django.contrib.auth.models import User


class TestVisualSchedulerPerformance:
    """Test Visual Scheduler performance with large datasets."""

    @pytest.fixture
    def large_dataset(self, db):
        """Create a large dataset for performance testing."""
        # This would need to be implemented with proper Django models
        # For now, we'll create a placeholder that can be extended
        pass

    @pytest.mark.slow
    def test_visual_scheduler_load_time(self, authenticated_page: Page, live_server_url: str):
        """Test Visual Scheduler loading performance with large dataset."""
        page = authenticated_page
        
        # Navigate to visual scheduler (this URL would need to be adjusted based on actual routing)
        # page.goto(f"{live_server_url}/admin/competition/season/1/visual-schedule/")
        
        # For now, we'll test navigation to the competitions section
        page.goto(f"{live_server_url}/admin/competition/")
        
        # Measure page load time
        with page.expect_response("**") as response_info:
            page.reload()
        
        response = response_info.value
        # Basic performance check - page should load in under 5 seconds
        # In a real implementation, we'd measure JavaScript execution time,
        # DOM rendering time, and drag-and-drop responsiveness
        
        assert response.status == 200

    @pytest.mark.slow 
    def test_drag_drop_performance(self, authenticated_page: Page, live_server_url: str):
        """Test drag and drop responsiveness with many elements."""
        page = authenticated_page
        
        # This test would:
        # 1. Create a season with many matches, fields, and time slots
        # 2. Open the Visual Scheduler
        # 3. Measure the time it takes to perform drag and drop operations
        # 4. Assert that operations complete within acceptable time limits
        
        # Placeholder implementation
        page.goto(f"{live_server_url}/admin/competition/")
        expect(page.locator("h1")).to_contain_text("Competition administration")

    @pytest.mark.slow
    def test_scheduler_memory_usage(self, authenticated_page: Page, live_server_url: str):
        """Test memory usage of Visual Scheduler with large datasets."""
        page = authenticated_page
        
        # This test would:
        # 1. Monitor browser memory usage
        # 2. Load Visual Scheduler with large dataset
        # 3. Perform various operations
        # 4. Check for memory leaks or excessive usage
        
        # Placeholder implementation
        page.goto(f"{live_server_url}/admin/competition/")
        
        # In a real implementation, we'd use Chrome DevTools Protocol
        # to monitor memory usage and performance metrics
        
        expect(page.locator("h1")).to_contain_text("Competition administration")