pyproject.toml
[project]
name = "vitriolic"
urls = { "Source Code" = "https://github.com/goodtune/vitriolic" }
description = "A suite of packages for the Django framework to provide integrated Content Management and Sports Administration functionality for sporting associations."
readme = "README.md"
license = "BSD-3-Clause"
authors = [
    { "name" = "Gary Reynolds", "email" = "gary@touch.asn.au" }
]
classifiers = []
requires-python = ">=3.11"
dependencies = [
    "django-froala-editor>=2.9",
    "django-guardian>=2.4",
    "django-modelforms",
    "django-mptt",
    "django-timezone-field",
    "django>=3.2",
    "first",
    "namedentities",
    "psycopg",
]
dynamic = ["version"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]
test = [
    "coverage",
    "django-environ",
    "django-test-plus",
    "factory_boy",
    "freezegun",
    "packaging>=25.0",
    "tblib",
]
e2e = [
    "playwright",
    "pytest-playwright",
    "pytest-django",
    "redis",
    "celery",
]

[project.optional-dependencies]
admin = ["django-bootstrap3", "django-gravatar2"]
common = []
content = []
news = [
    "babel",
    "django-imagekit",
    "python-dateutil",
    "python-magic",
]
competition = [
    "celery>5",
    "cloudinary",
    "django-embed-video",
    "django-formtools",
    "django-htmx>=1.23.2",
    "django-mathfilters",
    "djangorestframework>3.11",
    "drf-nested-routers",
    "google-api-python-client",
    "google-auth-oauthlib",
    "icalendar",
    "internationaltouch-oauth2client",
    "markdown",
    "pyparsing",
    "python-dateutil",
]
# Deprecated optional dependencies
redis = []

[build-system]
requires = ["setuptools", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
include = ["touchtechnology*", "tournamentcontrol*"]

[tool.setuptools_scm]

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
include_trailing_comma = true
multi_line_output = 5

[tool.black]
line-length = 88
target-version = ["py39", "py312"]
